+++ src/index.ts
@@ -1,3 +1,4 @@
+import { Scope } from '@sentry/node';
 import { ErrorRequestHandler } from 'express';
 import * as fs from 'fs';
 import isObject = require('lodash.isobject');
@@ -14,6 +15,7 @@ import { initLoggerStreams } from './streams';
 
 export type PinoLogger = pino.BaseLogger;
 export type Level = pino.LevelWithSilent;
+export type PinoHooks = { logMethod: (inputArgs: any, method: any) => void };
 
 export interface Cosmas extends PinoLogger {
     warning: pino.LogFn;
@@ -21,8 +23,28 @@ export interface Cosmas extends PinoLogger {
     express: CosmasExpressMiddleware;
     expressError: ErrorRequestHandler;
     stream: Writable;
+    realHooks: PinoHooks;
     (childName: string): Cosmas;
 }
+
+export interface LogFnSentry {
+    (msg: string, sentryCallback?: (scope: Scope) => void, ...args: any[]): void;
+    (obj: object, sentryCallback?: (scope: Scope) => void, ...args: any[]): void;
+    (obj: object, msg?: string, sentryCallback?: (scope: Scope) => void, ...args: any[]): void;
+}
+
+export interface CosmasSentry extends Cosmas {
+    fatal: LogFnSentry;
+    error: LogFnSentry;
+    warning: LogFnSentry;
+    warn: LogFnSentry;
+    info: LogFnSentry;
+    debug: LogFnSentry;
+    trace: LogFnSentry;
+    silent: LogFnSentry;
+}
+
+export * from './sentry';
 
 export interface CosmasFactory extends Cosmas {
     (data?: string | CosmasOptions): Cosmas;
@@ -52,7 +74,7 @@ const objEmpty = (obj: object) => Object.keys(obj).length === 0;
 // The original version was pino-multistream 4.2.0 (commit bf7941f) - https://github.com/pinojs/pino-multi-stream/blob/bf7941f77661b6c14dd40840ff4a4db6897f08eb/multistream.js#L43
 const maxLevelWrite: pino.WriteFn = function (this: any, data: object): void {
     let stream;
-    const metadata = Symbol.for('pino.metadata');
+    const metadata = pino.symbols.needsMetadataGsym;
     const level = this.lastLevel;
     const streams = this.streams;
     for (const dest of streams) {
@@ -105,30 +127,31 @@ const initFormatters = (options: CosmasOptions & { loggerName?: string }) => {
 };
 
 const initHooks = (options: CosmasOptions & { loggerName?: string }) => {
-    const hooks: { logMethod?: (inputArgs: any, method: any) => void } = {};
-    if (!options.loggerName) return hooks;
-
+    const realHooks: PinoHooks = {
+        logMethod(inputArgs, method) {
+            return method.apply(this, inputArgs);
+        },
+    };
+    const hooks: PinoHooks = {
+        logMethod(inputArgs, method) {
+            return realHooks.logMethod.call(this, inputArgs, method);
+        },
+    };
+    if (!options.loggerName) return { hooks, realHooks };
     // always put logger name to message
-    hooks.logMethod = function (inputArgs, method) {
+    realHooks.logMethod = function (inputArgs, method) {
         const text = inputArgs[inputArgs.length - 1];
         if (typeof text === 'string' || text instanceof String) {
             inputArgs[inputArgs.length - 1] = `[${options.loggerName}] ${text}`;
         }
         return method.apply(this, inputArgs);
     };
-    return hooks;
+    return { hooks, realHooks };
 };
 
 const defaultLogger = (options: CosmasOptions & { loggerName?: string } = {}): Cosmas => {
     serializers.disablePaths(options.disableFields);
     serializers.enablePaths(options.enableFields);
-
-    if (options.sentry) {
-        const sentry = require('@sentry/node');
-        if (typeof options.sentry === 'string') {
-            sentry.init({ dsn: options.sentry });
-        }
-    }
 
     const isTesting = process.env.NODE_ENV === 'test';
     const defaultLevel: Level = options.defaultLevel || (isTesting ? 'silent' : 'debug');
@@ -136,7 +159,7 @@ const defaultLogger = (options: CosmasOptions & { loggerName?: string } = {}): C
     const streams = initLoggerStreams(defaultLevel, Object.assign({}, options, { messageKey }));
 
     const formatters = initFormatters(options);
-    const hooks = initHooks(options);
+    const { hooks, realHooks } = initHooks(options);
 
     options.ignoredHttpMethods = options.ignoredHttpMethods || ['OPTIONS'];
     const logger = (pino(
@@ -161,7 +184,7 @@ const defaultLogger = (options: CosmasOptions & { loggerName?: string } = {}): C
 
     // Add maxLevel support to pino-multi-stream
     // This could be replaced with custom pass-through stream being passed to multistream, which would filter the messages
-    const loggerStream = (logger as any)[(pino as any).symbols.streamSym] as any;
+    const loggerStream = (logger as any)[pino.symbols.streamSym];
     const streamMaxLevelWrite = maxLevelWrite.bind(loggerStream);
     loggerStream.write = (chunk: any) => {
         streamMaxLevelWrite(chunk);
@@ -169,6 +192,7 @@ const defaultLogger = (options: CosmasOptions & { loggerName?: string } = {}): C
     };
     return Object.assign(logger, {
         options,
+        realHooks,
         express: expressMiddleware.bind(logger),
         expressError: expressErrorMiddleware as any,
     });
@@ -202,6 +226,6 @@ const loggerFactory = (data: string | CosmasOptions = {}, loggerOptions: CosmasO
     return loggerProxy;
 };
 
-const factoryProxy = makeCallable(loggerFactory(), loggerFactory);
+const factoryProxy: CosmasFactory = makeCallable(loggerFactory(), loggerFactory);
 
 export default factoryProxy;