+++ src/sentry.ts
@@ -1,7 +1,12 @@
-import { captureException, captureMessage, Severity, withScope } from '@sentry/node';
-import { Transform, TransformCallback } from 'stream';
-import { CosmasOptions } from './interfaces';
+import { captureException, captureMessage, Scope, Severity, withScope } from '@sentry/node';
+import { createNamespace } from 'cls-hooked';
+import * as pino from 'pino';
+import { Cosmas, CosmasSentry } from '.';
 import { levels } from './levels';
+
+const clsNamespace = createNamespace('cosmas.sentryExtend');
+
+type SentryCallback = (scope: Scope) => void;
 
 const reportToSentry = (obj: any) => {
     if (!obj.stack) {
@@ -23,20 +28,55 @@ const PINO_TO_SENTRY: { [key: number]: Severity } = {
     60: Severity.Critical,
 };
 
-export const createSentryTransformStream = (options: CosmasOptions): any => {
-    return class SentryTransformStream extends Transform {
-        // tslint:disable-next-line:function-name
-        public _transform(chunk: any, _encoding: string, callback: TransformCallback) {
-            const obj = JSON.parse(chunk);
-            if (obj.level >= (options.sentryLevel || levels.warn)) {
-                withScope((scope) => {
-                    scope.setLevel(PINO_TO_SENTRY[obj.level]);
-                    scope.setExtras(obj);
-                    reportToSentry(obj);
-                });
-            }
-            this.push(chunk);
-            callback();
-        }
+export const extendSentry = (logger: Cosmas, options: { sentry: string | true; sentryLevel?: number }) => {
+    const sentry = require('@sentry/node');
+    if (typeof options.sentry === 'string') {
+        sentry.init({ dsn: options.sentry });
+    }
+
+    const originalWrite = (logger as any)[pino.symbols.streamSym].write;
+    // unfortunately, this is the only place in pino, we can hook onto, where we can be sure all
+    // the hooks, formatters and serializers are already applied
+    (logger as any)[pino.symbols.streamSym].write = function (s: string) {
+        originalWrite.call(this, s);
+        const obj = JSON.parse(s);
+        if (obj.level < (options.sentryLevel || levels.warn)) return;
+        const sentryCallback: SentryCallback | undefined = clsNamespace.get('sentryCallback');
+        withScope((scope) => {
+            scope.setLevel(PINO_TO_SENTRY[obj.level]);
+            scope.setExtras(obj);
+            if (sentryCallback) sentryCallback(scope);
+            reportToSentry(obj);
+        });
     };
+
+    logger.realHooks.logMethod = function (inputArgs, method) {
+        // TODO: automatic types for logFn calls
+        let sentryCallback: SentryCallback | undefined;
+        let obj: object;
+        let rest: any;
+        let msg: string;
+        let newArgs: any;
+        if (typeof inputArgs[0] === 'string') {
+            [msg, sentryCallback, ...rest] = inputArgs;
+            newArgs = [msg, ...rest];
+        } else if (typeof inputArgs[1] === 'string') {
+            [obj, msg, sentryCallback, ...rest] = inputArgs;
+            newArgs = [obj, msg, ...rest];
+        } else {
+            [obj, sentryCallback, ...rest] = inputArgs;
+            newArgs = [obj, ...rest];
+        }
+
+        if (!sentryCallback) {
+            return method.apply(this, inputArgs);
+        }
+
+        clsNamespace.runAndReturn(() => {
+            clsNamespace.set('sentryCallback', sentryCallback);
+            return method.apply(this, newArgs);
+        });
+    };
+
+    return logger as CosmasSentry;
 };