+++ src/tests/sentry-mocked.test.ts
@@ -1,7 +1,10 @@
+import { Scope } from '@sentry/node';
 import omit = require('omit-deep');
+import { Cosmas, CosmasFactory, CosmasSentry } from '../index';
 import { levels } from '../levels';
 
-let loggerFactory;
+let loggerFactory: CosmasFactory;
+let extendSentry: (logger: Cosmas, options: { sentry: string | true; sentryLevel?: number }) => CosmasSentry;
 const scope: any = {};
 const withScope = jest.fn((fn) =>
     fn({
@@ -13,6 +16,9 @@ const withScope = jest.fn((fn) =>
         },
         setLevel: (level: any) => {
             scope.level = level;
+        },
+        setTags: (val: any) => {
+            scope.tags = val;
         },
     })
 );
@@ -44,6 +50,7 @@ describe('sentry mocked', () => {
             };
         });
         loggerFactory = require('../index').default;
+        extendSentry = require('../sentry').extendSentry;
     });
     beforeEach(() => {
         captureException.mockReset();
@@ -51,9 +58,9 @@ describe('sentry mocked', () => {
     });
     test('can create logger with options', () => {
         expect(() => loggerFactory()).not.toThrowError();
-        expect(() => loggerFactory({ sentry: true })).not.toThrowError();
+        expect(() => extendSentry(loggerFactory, { sentry: true })).not.toThrowError();
         expect(init).not.toHaveBeenCalled();
-        expect(() => loggerFactory({ sentry: 'dummy' })).not.toThrowError();
+        expect(() => extendSentry(loggerFactory, { sentry: 'dummy' })).not.toThrowError();
         expect(init.mock.calls[0]).toMatchInlineSnapshot(`
             Array [
               Object {
@@ -67,7 +74,8 @@ describe('sentry mocked', () => {
         const dateNow = Date.now;
         Date.now = jest.fn(() => 1520343036000);
         await new Promise((resolve, reject) => {
-            const logger = loggerFactory({
+            const logger = loggerFactory();
+            extendSentry(logger, {
                 sentry: 'DSN',
                 sentryLevel: levels.fatal,
             });
@@ -89,27 +97,26 @@ describe('sentry mocked', () => {
         `);
         expect(captureMessage.mock.results[0].value.scope.extras['cosmas.pkgVersion']).toBeDefined();
         expect(omit(captureMessage.mock.results[0].value, 'cosmas.pkgVersion')).toMatchInlineSnapshot(`
-Object {
-  "data": "fatal",
-  "scope": Object {
-    "extras": Object {
-      "level": 60,
-      "message": "fatal",
-      "severity": "CRITICAL",
-      "time": "2018-03-06T13:30:36.000Z",
-    },
-    "level": "critical",
-  },
-}
-`);
+            Object {
+              "data": "fatal",
+              "scope": Object {
+                "extras": Object {
+                  "level": 60,
+                  "message": "fatal",
+                  "severity": "CRITICAL",
+                  "time": "2018-03-06T13:30:36.000Z",
+                },
+                "level": "critical",
+              },
+            }
+        `);
         Date.now = dateNow;
     });
 
     test('sentry captureException with stack and correct levels', async () => {
         await new Promise((resolve, reject) => {
-            const logger = loggerFactory({
-                sentry: 'DSN',
-            });
+            const originalLogger = loggerFactory();
+            const logger = extendSentry(originalLogger, { sentry: 'DSN' });
             captureException.mockReset();
             captureException.mockImplementation(createCapture(resolve));
             logger.error(new Error());
@@ -122,5 +129,89 @@ Object {
                 level: 'error',
             },
         });
+    });
+
+    test('can pass sentry tags, context and extras', async () => {
+        const dateNow = Date.now;
+        Date.now = jest.fn(() => 1520343036000);
+        await new Promise((resolve, reject) => {
+            const originalLogger = loggerFactory();
+            const logger = extendSentry(originalLogger, { sentry: 'DSN' });
+            captureMessage.mockImplementation(createCapture(resolve));
+            logger.fatal({ name: 'John Doe' }, 'sentryData', (scope: Scope) => {
+                scope.setContext('dummyContext', { foo: 'bar' });
+                scope.setTags({ first: 'firstTag' });
+                scope.setExtras({ extra: 'extraValue' });
+            });
+        });
+        expect(captureMessage).toHaveBeenCalledTimes(1);
+        expect(captureException).not.toHaveBeenCalled();
+        expect(captureMessage.mock.calls[0]).toMatchInlineSnapshot(`
+            Array [
+              "sentryData",
+            ]
+        `);
+        expect(omit(captureMessage.mock.results[0].value, 'cosmas.pkgVersion')).toMatchInlineSnapshot(`
+            Object {
+              "data": "sentryData",
+              "scope": Object {
+                "context": Object {
+                  "dummyContext": Object {
+                    "foo": "bar",
+                  },
+                },
+                "extras": Object {
+                  "extra": "extraValue",
+                },
+                "level": "critical",
+                "tags": Object {
+                  "first": "firstTag",
+                },
+              },
+            }
+        `);
+        Date.now = dateNow;
+    });
+
+    test('can pass sentry tags, context and extras with 3 params', async () => {
+        const dateNow = Date.now;
+        Date.now = jest.fn(() => 1520343036000);
+        await new Promise((resolve, reject) => {
+            const originalLogger = loggerFactory();
+            const logger = extendSentry(originalLogger, { sentry: 'DSN' });
+            captureMessage.mockImplementation(createCapture(resolve));
+            logger.fatal('sentryfatal', (scope: Scope) => {
+                scope.setContext('dummyContext', { foo: 'bar' });
+                scope.setTags({ first: 'firstTag' });
+                scope.setExtras({ extra: 'extraValue' });
+            });
+        });
+        expect(captureMessage).toHaveBeenCalledTimes(1);
+        expect(captureException).not.toHaveBeenCalled();
+        expect(captureMessage.mock.calls[0]).toMatchInlineSnapshot(`
+            Array [
+              "sentryfatal",
+            ]
+        `);
+        expect(omit(captureMessage.mock.results[0].value, 'cosmas.pkgVersion')).toMatchInlineSnapshot(`
+            Object {
+              "data": "sentryfatal",
+              "scope": Object {
+                "context": Object {
+                  "dummyContext": Object {
+                    "foo": "bar",
+                  },
+                },
+                "extras": Object {
+                  "extra": "extraValue",
+                },
+                "level": "critical",
+                "tags": Object {
+                  "first": "firstTag",
+                },
+              },
+            }
+        `);
+        Date.now = dateNow;
     });
 });